+++ src/OpenTelemetry/CHANGELOG.md
@@ -9,7 +9,7 @@ please check the latest changes
 
 ## Unreleased
 
-* Added `TracerProviderOptions` and `SetErrorStatusOnUnhandledException`.
+* Added `TracerProviderOptions` and `SetErrorStatusOnException`.
   ([#1858](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1858))
 * Added `ForceFlush` to `TracerProvider`.
   ([#1837](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1837))